\name{MixedPoisson2-package}
\alias{MixedPoisson2}
\docType{package}
\title{
Mixed Poisson models
}
\description{
The package provides functions fitting parameters of different mixed Poisson models. As the method of estimation Expectation-Maximization (EM) algorithm is used.}
\details{
\tabular{ll}{
Package: \tab MixedPoisson\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-13\cr
License: \tab GPL-2\cr
}

}
\author{
Alicja Wolny-Dominiak and Michal Trzesiok

Maintainer: <alicja.wolny-dominiak@ue.katowice.pl>

}
\references{
Karlis, D. (2005). EM algorithm for mixed Poisson and other discrete distributions. Astin Bulletin, 35(01), 3-24.
Ghitany, M. E., Karlis, D., Al-Mutairi, D. K., & Al-Awadhi, F. A. (2012). An EM algorithm for multivariate mixed Poisson regression models and its application. Applied Mathematical Sciences, 6(137), 6843-6856.
}
