\name{MixedPoisson-package}
\alias{MixedPoisson-package}
\alias{MixedPoisson}
\docType{package}
\title{
Mixed Poisson Models
}
\description{
The package provides functions for estimating parameters of different mixed Poisson models. Expectation-Maximization (EM) algorithm is used as the estimation technique.}
\details{
\tabular{ll}{
Package: \tab MixedPoisson\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-29\cr
License: \tab GPL-2\cr
}

}
\author{
Alicja Wolny-Dominiak and Michal Trzesiok

Maintainer: <alicja.wolny-dominiak@ue.katowice.pl>

}
\references{
Karlis, D. (2005). EM algorithm for mixed Poisson and other discrete distributions. Astin bulletin, 35(01), 3-24.
Ghitany, M. E., Karlis, D., Al-Mutairi, D. K., & Al-Awadhi, F. A. (2012). An EM algorithm for multivariate mixed Poisson regression models and its application. Applied Mathematical Sciences, 6(137), 6843-6856.
}
